Encapsulate a method in private scope
There is a method NightWriter#lookup that is only used within the NightWriter class itself. This commit encapsulates that method in private scope and removes the unit tests covering it, since those tests rely on calling it from outside the class definition. The behavior is still tested by manner of the testing on the public methods that call the internals.